Backend: In Appointments.java, I added a @JsonIgnore to the 'data' field so that it doesn't return a pet image with each Appointment record when a GET is called
Frontend: Made the pet images from getAllPets() reusable by storing them in a hook in Dashboard.js. The Pet Images are then passed down to Dashboard Appointments where they can be used in the UpcomingAppointments and PastAppointments tables. This eliminates the need for using the old getAllAppointments() endpoint that returns a pet image with each Appointment record.